requirements.txt
pyyaml
jsonschema
hvac
boto3
google-cloud-secret-manager
nats-py
ampy-bus
structlog~=24.4
python-json-logger~=2.0
opentelemetry-api~=1.27
opentelemetry-sdk~=1.27
opentelemetry-exporter-otlp-proto-http~=1.27
prometheus-client